Allow non-prefix files if prefix version missing
This fixes #5

== modules

Prefixed files are checked first and then fallback to non-prefixed
files.

== globs

All the matches are checked and if a prefixed version can't be found but
a non-prefixed version exists that match is used instead. Example:

--- _baz.css
--- bar.css
--- baz.css

Given the above list of files and `*.css` as the glob it would return
[`_baz.css`, `bar.css`]